constArg=ConstantInt::get(IntegerType::get(V->getContext(),32), (uint64_t)V);

+++ sample/lib/Hello/Hello.cpp
@@ -61,7 +61,7 @@ struct Hello :  public FunctionPass
 								}
 							if (areArgsConst)
 								{
-								constArg=ConstantInt::get(IntegerType::get(V->getContext(),32), V);
+								constArg=ConstantInt::get(IntegerType::get(V->getContext(),32), (uint64_t)V);
 								/*calledFunc=callInst->getCalledFunction();
 								errs()<< "Function ";
 								errs().write_escaped(calledFunc->getName())<<  " is called with actual arguments ";
